requirements.txt
aiogram~=2.14
environs~=8.0.0
python-dotenv~=0.21.1